reorganizing + docstrings
