+++ eegdash/features/decorators.py
@@ -17,19 +17,23 @@ The module provides the following decorators:
 
 import functools
 import inspect
+import os
 from collections.abc import Callable
 from typing import Iterable, List, Tuple, Type
 
-from .extractors import (
-    BasePreprocessorOutputType,
-    BivariateFeature,
+from .base_utils import (
     BivariateIterator,
-    DirectedBivariateFeature,
+    channel_names_to_indices,
+    get_underlying_func,
+)
+from .kinds import (
+    BivariateFeature,
     MultivariateFeature,
     UnivariateFeature,
-    _get_underlying_func,
 )
-from .utils import channel_names_to_indices
+from .output_types import (
+    BasePreprocessorOutputType,
+)
 
 __all__ = [
     "bivariate_feature",
@@ -47,6 +51,7 @@ _WRAPPER_ASSIGNMENTS = [
     "feature_kind",
     "metadata_preprocessor",
 ]
+SPHINX_BUILD = bool(os.environ.get("SPHINX_BUILD", ""))
 
 
 def _update_wrapper(
@@ -61,12 +66,14 @@ def _update_wrapper(
     wrapped is the original function
     assigned is a tuple naming the attributes assigned directly
     from the wrapped function to the wrapper function (defaults to
-    functools.WRAPPER_ASSIGNMENTS)
+    eegdash.features.decorators._WRAPPER_ASSIGNMENTS)
     updated is a tuple naming the attributes of the wrapper that
     are updated with the corresponding attribute from the wrapped
     function (defaults to functools.WRAPPER_UPDATES)
     """
-    wrapped_f = _get_underlying_func(wrapped)
+    if SPHINX_BUILD:
+        return wrapped  # fool sphinx
+    wrapped_f = get_underlying_func(wrapped)
     for attr in assigned:
         try:
             value = getattr(wrapped_f, attr)
@@ -76,6 +83,7 @@ def _update_wrapper(
             setattr(wrapper, attr, value)
     for attr in updated:
         getattr(wrapper, attr).update(getattr(wrapped_f, attr, {}))
+    wrapper.__signature__ = inspect.signature(wrapped)
     # Issue #17482: set __wrapped__ last so we don't inadvertently copy it
     # from the wrapped function when updating __dict__
     wrapper.__wrapped__ = wrapped
@@ -146,7 +154,7 @@ class FeaturePredecessor:
             attribute attached.
 
         """
-        f = _get_underlying_func(func)
+        f = get_underlying_func(func)
         f.parent_extractor_type = self.parent_extractor_type
         return func
 
@@ -189,7 +197,7 @@ class FeatureKind:
             The decorated function with the `feature_kind` attribute set.
 
         """
-        f = _get_underlying_func(func)
+        f = get_underlying_func(func)
         f.feature_kind = self.feature_kind
         return func
 
@@ -204,33 +212,14 @@ original channel names.
 """
 
 
-def bivariate_feature(func: Callable, directed: bool = False) -> Callable:
-    r"""Decorator to mark a feature as bivariate.
+bivariate_feature = FeatureKind(BivariateFeature())
+r"""Decorator to mark a feature as bivariate.
 
-    Specifies that the feature operates on pairs of channels.
+Specifies that the feature operates on pairs of channels.
+The output will be formatted as a dictionary with keys matching the
+original channel name pairs.
 
-    Parameters
-    ----------
-    func : callable
-        The feature extraction function to decorate.
-    directed : bool, default False
-        If True, the feature is treated as directed.
-        If False, only unique, unordered pairs are computed.
-
-    Returns
-    -------
-    callable
-        The decorated function with either a
-        :class:`~eegdash.features.extractors.BivariateFeature` or
-        :class:`~eegdash.features.extractors.DirectedBivariateFeature`
-        kind attached.
-
-    """
-    if directed:
-        kind = DirectedBivariateFeature()
-    else:
-        kind = BivariateFeature()
-    return FeatureKind(kind)(func)
+"""
 
 
 multivariate_feature = FeatureKind(MultivariateFeature())
@@ -284,19 +273,18 @@ class PreprocessorOutputType:
             `output_type`, with the original preprocessor function as its implementation.
 
         """
-        return _wraps(preprocessor)(
-            type(
-                preprocessor.__name__,
-                (self.output_type,),
-                {
-                    "__call__": self.output_type._call_metadata
-                    if "_metadata" in inspect.signature(preprocessor).parameters
-                    else self.output_type._call,
-                    "__module__": preprocessor.__module__,
-                    "__doc__": preprocessor.__doc__,
-                },
-            )(preprocessor)
+        preprocessor_class = type(
+            preprocessor.__name__,
+            (self.output_type,),
+            {
+                "__call__": self.output_type._call_metadata
+                if "_metadata" in inspect.signature(preprocessor).parameters
+                else self.output_type._call,
+            },
         )
+        preprocessor_instance = preprocessor_class(preprocessor)
+        preprocessor_instance = _update_wrapper(preprocessor_instance, preprocessor)
+        return preprocessor_instance
 
 
 def metadata_preprocessor(func: Callable):
@@ -316,7 +304,7 @@ def metadata_preprocessor(func: Callable):
         The decorated function with the `metadata_preprocessor` attribute set.
 
     """
-    f = _get_underlying_func(func)
+    f = get_underlying_func(func)
     if "_metadata" not in inspect.signature(f).parameters:
         raise TypeError(
             f"{f.__name__} cannot be set as a metadata preprocessor "
@@ -328,6 +316,21 @@ def metadata_preprocessor(func: Callable):
 
 
 class ChannelPairer:
+    r"""Decorator to set a feature preprocessor as a channel pairer.
+
+    This decorator lets a feature preprocessor get an additional ``pairs``
+    keyword argument, and sets a metadata field named ``'ch_pair_iterator'``
+    containing a :class:`~eegdash.features.base_utils.BivariateIterator`
+    accordingly before calling the underlying preprocessor.
+
+    Parameters
+    ----------
+    directed : bool
+        Whether the preprocessor assumes *directed* or *undirected* bivariate
+        iteration.
+
+    """
+
     def __init__(self, directed: bool = False):
         self.directed = directed
 
@@ -350,7 +353,7 @@ class ChannelPairer:
             _metadata["ch_pair_iterator"] = BivariateIterator(
                 pairs, directed=self.directed
             )
-            f = _get_underlying_func(func)
+            f = get_underlying_func(func)
             if "_metadata" in inspect.signature(f).parameters:
                 kwargs["_metadata"] = _metadata
             if hasattr(f, "metadata_preprocessor") and f.metadata_preprocessor:
@@ -363,4 +366,17 @@ class ChannelPairer:
 
 # Syntax sugar
 channel_pairer = ChannelPairer(directed=False)
+r"""Decorator to mark a feature preprocessor as an undirected channel pairer.
+
+Specifies that the feature preprocessor operates on undirected pairs of
+channels.
+
+"""
+
 channel_directed_pairer = ChannelPairer(directed=True)
+r"""Decorator to mark a feature preprocessor as an undirected channel pairer.
+
+Specifies that the feature preprocessor operates on undirected pairs of
+channels.
+
+"""
